schemaexpr: move more from sqlbase and break the dependency cycle
Release note: None
