+++ pkg/sql/sqlbase/table.go
@@ -21,67 +21,13 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/sql/catalog/descpb"
 	"github.com/cockroachdb/cockroach/pkg/sql/pgwire/pgcode"
 	"github.com/cockroachdb/cockroach/pkg/sql/pgwire/pgerror"
+	"github.com/cockroachdb/cockroach/pkg/sql/schemaexpr"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
 	"github.com/cockroachdb/cockroach/pkg/sql/types"
 	"github.com/cockroachdb/cockroach/pkg/util/hlc"
 	"github.com/cockroachdb/cockroach/pkg/util/log"
 	"github.com/cockroachdb/errors"
 )
-
-// SanitizeVarFreeExpr verifies that an expression is valid, has the correct
-// type and contains no variable expressions. It returns the type-checked and
-// constant-folded expression.
-func SanitizeVarFreeExpr(
-	ctx context.Context,
-	expr tree.Expr,
-	expectedType *types.T,
-	context string,
-	semaCtx *tree.SemaContext,
-	maxVolatility tree.Volatility,
-) (tree.TypedExpr, error) {
-	if tree.ContainsVars(expr) {
-		return nil, pgerror.Newf(pgcode.Syntax,
-			"variable sub-expressions are not allowed in %s", context)
-	}
-
-	// We need to save and restore the previous value of the field in
-	// semaCtx in case we are recursively called from another context
-	// which uses the properties field.
-	defer semaCtx.Properties.Restore(semaCtx.Properties)
-
-	// Ensure that the expression doesn't contain special functions.
-	flags := tree.RejectSpecial
-
-	switch maxVolatility {
-	case tree.VolatilityImmutable:
-		flags |= tree.RejectStableOperators
-		fallthrough
-
-	case tree.VolatilityStable:
-		flags |= tree.RejectVolatileFunctions
-
-	case tree.VolatilityVolatile:
-		// Allow anything (no flags needed).
-
-	default:
-		panic(errors.AssertionFailedf("maxVolatility %s not supported", maxVolatility))
-	}
-	semaCtx.Properties.Require(context, flags)
-
-	typedExpr, err := tree.TypeCheck(ctx, expr, semaCtx, expectedType)
-	if err != nil {
-		return nil, err
-	}
-
-	actualType := typedExpr.ResolvedType()
-	if !expectedType.Equivalent(actualType) && typedExpr != tree.DNull {
-		// The expression must match the column type exactly unless it is a constant
-		// NULL value.
-		return nil, fmt.Errorf("expected %s expression to have type %s, but '%s' has type %s",
-			context, expectedType, expr, actualType)
-	}
-	return typedExpr, nil
-}
 
 // MakeColumnDefDescs creates the column descriptor for a column, as well as the
 // index descriptor if the column is a primary key or unique.
@@ -139,7 +85,7 @@ func MakeColumnDefDescs(
 		// Verify the default expression type is compatible with the column type
 		// and does not contain invalid functions.
 		var err error
-		if typedExpr, err = SanitizeVarFreeExpr(
+		if typedExpr, err = schemaexpr.SanitizeVarFreeExpr(
 			ctx, d.DefaultExpr.Expr, resType, "DEFAULT", semaCtx, tree.VolatilityVolatile,
 		); err != nil {
 			return nil, nil, nil, err
@@ -200,7 +146,7 @@ func EvalShardBucketCount(
 	ctx context.Context, semaCtx *tree.SemaContext, evalCtx *tree.EvalContext, shardBuckets tree.Expr,
 ) (int32, error) {
 	const invalidBucketCountMsg = `BUCKET_COUNT must be an integer greater than 1`
-	typedExpr, err := SanitizeVarFreeExpr(
+	typedExpr, err := schemaexpr.SanitizeVarFreeExpr(
 		ctx, shardBuckets, types.Int, "BUCKET_COUNT", semaCtx, tree.VolatilityVolatile,
 	)
 	if err != nil {
